+++ rotkehlchen/data_import/manager.py
@@ -4,6 +4,7 @@ from typing import Any
 from rotkehlchen.data_import.importers.binance import BinanceImporter
 from rotkehlchen.data_import.importers.bisq_trades import BisqTradesImporter
 from rotkehlchen.data_import.importers.bitcoin_tax import BitcoinTaxImporter
+from rotkehlchen.data_import.importers.bitmex import BitMEXImporter
 from rotkehlchen.data_import.importers.blockfi_trades import BlockfiTradesImporter
 from rotkehlchen.data_import.importers.blockfi_transactions import BlockfiTransactionsImporter
 from rotkehlchen.data_import.importers.cointracking import CointrackingImporter
@@ -31,6 +32,7 @@ class DataImportSource(SerializableEnumNameMixin):
     ROTKI_TRADES = 10
     ROTKI_EVENTS = 11
     BITCOIN_TAX = 12
+    BITMEX_WALLET_HISTORY = 13
 
     def get_importer_type(self) -> type[BaseExchangeImporter]:
         if self == DataImportSource.COINTRACKING:
@@ -57,6 +59,8 @@ class DataImportSource(SerializableEnumNameMixin):
             return RotkiGenericEventsImporter
         if self == DataImportSource.BITCOIN_TAX:
             return BitcoinTaxImporter
+        if self == DataImportSource.BITMEX_WALLET_HISTORY:
+            return BitMEXImporter
         raise AssertionError(f'Unknown DataImportSource value {self}')