+++ rotkehlchen/tests/utils/dataimport.py
@@ -37,7 +37,7 @@ from rotkehlchen.db.filtering import (
 )
 from rotkehlchen.db.history_events import DBHistoryEvents
 from rotkehlchen.db.ledger_actions import DBLedgerActions
-from rotkehlchen.exchanges.data_structures import AssetMovement, Trade
+from rotkehlchen.exchanges.data_structures import AssetMovement, MarginPosition, Trade
 from rotkehlchen.fval import FVal
 from rotkehlchen.rotkehlchen import Rotkehlchen
 from rotkehlchen.tests.utils.constants import A_AXS, A_CRO, A_GBP, A_MCO, A_XMR
@@ -1210,6 +1210,127 @@ def assert_bisq_trades_import_results(rotki: Rotkehlchen):
         notes='ID: xxhee',
     )]
     assert trades == expected_trades
+
+
+def assert_bitmex_import_wallet_history(rotki: Rotkehlchen):
+    expected_asset_movements = [
+        AssetMovement(
+            location=Location.BITMEX,
+            category=AssetMovementCategory.DEPOSIT,
+            address=None,
+            transaction_id=None,
+            timestamp=Timestamp(1574825791),
+            asset=A_BTC,
+            amount=FVal(0.05000000),
+            fee_asset=A_BTC,
+            fee=Fee(FVal(00000000)),
+            link='Imported from BitMEX CSV file. Transact Type: Deposit',
+        ),
+        AssetMovement(
+            location=Location.BITMEX,
+            category=AssetMovementCategory.WITHDRAWAL,
+            address='3Qsy5NGSnGA1vd1cmcNgeMjLrKPsKNhfCe',
+            transaction_id=None,
+            timestamp=Timestamp(1577252845),
+            asset=A_BTC,
+            amount=FVal(0.05746216),
+            fee_asset=A_BTC,
+            fee=Fee(FVal(0.00100000)),
+            link='Imported from BitMEX CSV file. Transact Type: Withdrawal',
+        ),
+    ]
+    expected_margin_positions = [
+        MarginPosition(
+            location=Location.BITMEX,
+            open_time=None,
+            close_time=Timestamp(1576738800),
+            profit_loss=AssetAmount(FVal(0.00000373)),
+            pl_currency=A_BTC,
+            fee=Fee(FVal(0)),
+            fee_currency=A_BTC,
+            link='Imported from BitMEX CSV file. Transact Type: RealisedPNL',
+            notes='PnL from trade on XBTUSD',
+        ),
+        MarginPosition(
+            location=Location.BITMEX,
+            open_time=None,
+            close_time=Timestamp(1576825200),
+            profit_loss=AssetAmount(FVal(0.00000016)),
+            pl_currency=A_BTC,
+            fee=Fee(FVal(0)),
+            fee_currency=A_BTC,
+            link='Imported from BitMEX CSV file. Transact Type: RealisedPNL',
+            notes='PnL from trade on XBTUSD',
+        ),
+        MarginPosition(
+            location=Location.BITMEX,
+            open_time=None,
+            close_time=Timestamp(1576911600),
+            profit_loss=AssetAmount(FVal(-0.00000123)),
+            pl_currency=A_BTC,
+            fee=Fee(FVal(0)),
+            fee_currency=A_BTC,
+            link='Imported from BitMEX CSV file. Transact Type: RealisedPNL',
+            notes='PnL from trade on XBTUSD',
+        ),
+        MarginPosition(
+            location=Location.BITMEX,
+            open_time=None,
+            close_time=Timestamp(1576998000),
+            profit_loss=AssetAmount(FVal(-0.00000075)),
+            pl_currency=A_BTC,
+            fee=Fee(FVal(0)),
+            fee_currency=A_BTC,
+            link='Imported from BitMEX CSV file. Transact Type: RealisedPNL',
+            notes='PnL from trade on XBTUSD',
+        ),
+        MarginPosition(
+            location=Location.BITMEX,
+            open_time=None,
+            close_time=Timestamp(1577084400),
+            profit_loss=AssetAmount(FVal(-0.00000203)),
+            pl_currency=A_BTC,
+            fee=Fee(FVal(0)),
+            fee_currency=A_BTC,
+            link='Imported from BitMEX CSV file. Transact Type: RealisedPNL',
+            notes='PnL from trade on XBTUSD',
+        ),
+        MarginPosition(
+            location=Location.BITMEX,
+            open_time=None,
+            close_time=Timestamp(1577170800),
+            profit_loss=AssetAmount(FVal(-0.00000201)),
+            pl_currency=A_BTC,
+            fee=Fee(FVal(0)),
+            fee_currency=A_BTC,
+            link='Imported from BitMEX CSV file. Transact Type: RealisedPNL',
+            notes='PnL from trade on XBTUSD',
+        ),
+        MarginPosition(
+            location=Location.BITMEX,
+            open_time=None,
+            close_time=Timestamp(1577257200),
+            profit_loss=AssetAmount(FVal(0.00085517)),
+            pl_currency=A_BTC,
+            fee=Fee(FVal(0)),
+            fee_currency=A_BTC,
+            link='Imported from BitMEX CSV file. Transact Type: RealisedPNL',
+            notes='PnL from trade on XBTUSD',
+        ),
+    ]
+    with rotki.data.db.conn.read_ctx() as cursor:
+        margin_positions = rotki.data.db.get_margin_positions(cursor)
+        warnings = rotki.msg_aggregator.consume_warnings()
+        errors = rotki.msg_aggregator.consume_errors()
+        asset_movements = rotki.data.db.get_asset_movements(
+            cursor,
+            filter_query=AssetMovementsFilterQuery.make(),
+            has_premium=True,
+        )
+    assert asset_movements == expected_asset_movements
+    assert margin_positions == expected_margin_positions
+    assert len(warnings) == 0
+    assert len(errors) == 0
 
 
 def assert_binance_import_results(rotki: Rotkehlchen):